Merge git ignore lists from root- and subfolders
In 2016 and 2017 nobody copied the .gitignore file from the 2014 or 2015
folder. This will be probably forgotten again in the future for the next
MV. So put all patterns to the root .gitignore file.

This should also prevent committing a mv.pdf which had to be corrected
twice in the past as you can see in #2 and #8, and force you to put the
resulting pdf in the designated folder as discussed in #2.